+++ meta_package_manager/tests/test_cli_search.py
@@ -90,12 +90,21 @@ class TestSearch(CLISubCommandTests, CLITableTests):
         assert "ubersicht" in result.stdout
         assert "Übersicht" not in result.stdout
 
-    # PyPi's online search API is rate-limited. We add an artificial 2-seconds delay to
-    # prevent this issue:
+    # PyPi's online search API was at first rate-limited. So we added an artificial
+    # 2-seconds delay to prevent the following error:
     #   xmlrpc.client.Fault: <Fault -32500: 'HTTPTooManyRequests: The action could not
-    #   be performed because there were too many requests by the client. Limit may reset
-    #   in 1 seconds.'>
+    #   be performed because there were too many requests by the client. Limit may
+    #   reset in 1 seconds.'>
+    # Then the search API was shutdown altogether as it was hammered (see
+    # https://github.com/pypa/pip/issues/5216#issuecomment-744605466) which produced
+    # this error:
+    #   xmlrpc.client.Fault: <Fault -32500: "RuntimeError: PyPI's XMLRPC API has been
+    #   temporarily disabled due to unmanageable load and will be deprecated in the
+    #   near future. See https://status.python.org/ for more information.">
 
+    skip_pip_search = pytest.mark.skip(reason="pip search is deprecated")
+
+    @skip_pip_search
     def test_exact_search_tokenizer_one_result(self, invoke):
         sleep(2)
         result = invoke("--manager", "pip", "search", "--exact", "sed", color=False)
@@ -103,6 +112,7 @@ class TestSearch(CLISubCommandTests, CLITableTests):
         assert "1 package total" in result.stdout
         assert " sed " in result.stdout
 
+    @skip_pip_search
     @pytest.mark.parametrize(
         "query", ["SED", "SeD", "sEd*", "*sED*", "_seD-@", "", "_"]
     )
@@ -113,6 +123,7 @@ class TestSearch(CLISubCommandTests, CLITableTests):
         assert "0 package total" in result.stdout
         assert "sed" not in result.stdout
 
+    @skip_pip_search
     @pytest.mark.parametrize("query", ["", "_", "_seD-@"])
     def test_fuzzy_search_tokenizer_no_results(self, invoke, query):
         sleep(2)
@@ -121,6 +132,7 @@ class TestSearch(CLISubCommandTests, CLITableTests):
         assert "0 package total" in result.stdout
         assert "sed" not in result.stdout
 
+    @skip_pip_search
     @pytest.mark.parametrize("query", ["sed", "SED", "SeD", "sEd*", "*sED*"])
     def test_fuzzy_search_tokenizer_multiple_results(self, invoke, query):
         sleep(2)
@@ -130,6 +142,7 @@ class TestSearch(CLISubCommandTests, CLITableTests):
         assert " sed " in result.stdout
         assert " SED-cli " in result.stdout
 
+    @skip_pip_search
     @pytest.mark.parametrize("query", ["", "_", "_seD-@"])
     def test_extended_search_tokenizeri_no_results(self, invoke, query):
         sleep(2)
@@ -138,6 +151,7 @@ class TestSearch(CLISubCommandTests, CLITableTests):
         assert "0 package total" in result.stdout
         assert "sed" not in result.stdout
 
+    @skip_pip_search
     @pytest.mark.parametrize("query", ["sed", "SED", "SeD", "sEd*", "*sED*"])
     def test_extended_search_tokenizeri_multiple_results(self, invoke, query):
         sleep(2)